Add `NoArgs` validation to all of the commands
All the commands receive their arguments by passing flags and their values. Therefore, we would like to verify that passing arguments in any other way is not allowed, to avoid unwanted behavior. For example, in the `add workload` command we received unwanted behavior as described in #77.

This PR includes adding: `Args: cobra.NoArgs` to all of the commands, which will verify the above.

Signed-off-by: arielireni <aireni@redhat.com>