Bug fix: change `certsMap` to a local variable
This PR fixes the following bug in the `add deviceset` command:
```
→ ./bin/flotta add deviceset -n set1 -s 2
device-set 'set1' was added
device 'set1-device1' was added successfully to device-set 'set1' (1/2)
NewDeviceToSet failed: cannot copy certificates to device: open /tmp/certs378247597/ca.pem: no such file or directory. Device: set1-device2
```
The error occurred since the path of the certs directory of the first device is still in `certsMap` (since it was a global variable), but the path does not exist in the container of the second device.

Signed-off-by: arielireni <aireni@redhat.com>